requirements.txt
asgiref==3.2.7
Django==3.0.5
pytz==2019.3
sqlparse==0.3.1

azure-common==1.1.16
azure-core==1.3.0
azure-mgmt-nspkg==3.0.2
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-nspkg==3.0.0
django-storages==1.9.1